requirements.txt
django
crispy_bootstrap4
Pillow
django_filter
djangorestframework